Restructured code base
Moved lexer related functionality in lexer package and added parser package for parser relatet functionality
